trying to do jwt_required only for post route

+++ app/models/twiglet.py
@@ -4,8 +4,8 @@ import datetime
 
 class Twiglet(db.Model):
     twiglet_id = db.Column(db.Integer, primary_key=True)
-    longitude = db.Column(db.Integer, nullable=False)
-    latitude = db.Column(db.Integer, nullable=False)
+    longitude = db.Column(db.Float, nullable=False)
+    latitude = db.Column(db.Float, nullable=False)
     shop_name = db.Column(db.String)
     address = db.Column(db.String)
     found_by_user = db.Column(db.Integer, db.ForeignKey("user.user_id"))
